% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coviddata.R
\docType{data}
\name{coviddata}
\alias{coviddata}
\title{coviddata}
\format{
A list with two components.
\describe{
\item{data}{A data frame with four columns, date, location_code, cases and pop}
\item{adj}{A 11x11 matrix giving the adjency matrix between the 11 counties in Norway}
}
}
\source{
\url{https://github.com/folkehelseinstituttet/surveillance_data/tree/master/covid19}
}
\usage{
coviddata
}
\description{
Test data for Covid-18 extracted form the Norwegian institute of public health
The data have been extracted by the script Make_data_covid_test.R.
See the scripts Covid_GCF.R and Covid_SC.R on how to use these data
}
\keyword{datasets}
