% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_projector.R
\name{create_projector_matrix_interactionIV}
\alias{create_projector_matrix_interactionIV}
\title{Create the projector matrix fast
This is even fast for grid sizes of 40 000!}
\usage{
create_projector_matrix_interactionIV(ns, nt)
}
\description{
Create the projector matrix fast
This is even fast for grid sizes of 40 000!
}
