% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRF_RW.R
\name{GMRF_RW}
\alias{GMRF_RW}
\title{Creates the structure matrix for random walk order 1 (RW1) and RW2
The precision matrix is not scaled}
\usage{
GMRF_RW(n = 10, order = 1)
}
\arguments{
\item{n}{gives the number of time steps}

\item{order}{gives the order of the random walk}
}
\value{
a matrix with the precision matrix
}
\description{
Creates the structure matrix for random walk order 1 (RW1) and RW2
The precision matrix is not scaled
}
