% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikCorrectGeneric0.R
\name{LikCorrectGeneric0}
\alias{LikCorrectGeneric0}
\title{Calculated the correction factor needed for the generic0 model within INLA}
\usage{
LikCorrectGeneric0(Q, A, eps)
}
\arguments{
\item{'Q'}{The precision matrix}

\item{'A'}{The constraint matrix}

\item{'eps'}{The noise terms added to the diagonal}
}
\value{
The correction factor (for the LOG-marginal likelohood)
}
\description{
Calculated the correction factor needed for the generic0 model within INLA
}
