% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymodel.R
\name{mymodel}
\alias{mymodel}
\title{This function fits the model as done by Rachel Lowe (2021).
NB: INLA mode experimental is used.}
\usage{
mymodel(
  formula,
  data = df,
  family = "nbinomial",
  config = FALSE,
  num.threads = num.threads
)
}
\arguments{
\item{formula}{is the formula in INLA format}

\item{family}{is the distribution family fitted, e.g Poisson, Gaussian}

\item{df}{is the data to be used in the model fitting}
}
\value{

}
\description{
This function fits the model as done by Rachel Lowe (2021).
NB: INLA mode experimental is used.
}
