% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceTimeProjConstr.R
\name{SpaceTimeProjConstr}
\alias{SpaceTimeProjConstr}
\title{Create the projector matrix and the corresponding constraints}
\usage{
SpaceTimeProjConstr(ns, nt, method = "DEFAULT")
}
\arguments{
\item{'ns'}{The number of spatial points}

\item{'nt'}{The number of temporal points}

\item{'method'}{Either 'space' in which case the constraints with respect to space are used, or
'time' in which case the constraints with respect to time is used. If no specification is given,
'space' is given if ns>nt and 'time' in the opposite case.}
}
\value{
A list containing the projection matrix 'P' and the remaining constraints 'C', both of class 'dgCMatrix'.
}
\description{
Create the projector matrix and the corresponding constraints
}
