% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatioTemporalDataset.R
\docType{data}
\name{SpatioTemporalDataset}
\alias{SpatioTemporalDataset}
\title{SpatioTemporalDataset}
\format{
A data.table with 4 coloumns
\describe{
\item{main_spatial}{Index for the spatial effect}
\item{main_temporal}{Index for the temporal effect}
\item{interaction}{Index for the spatio-temporal effect}
\item{main_spatial}{The responce variable}
}
}
\source{
\url{https://github.com/folkehelseinstituttet/surveillance_data/tree/master/covid19}
}
\usage{
SpatioTemporalDataset
}
\description{
Simulated data set using the routine simulate_data.R
Data are simulated fron the Knorr-Held type IV model with the spatial model defined
through the spatial structure of the 54 regions in Germany and the temporal effect
following i RW(2) model.
The response is generated from a Poisson model with intercept equal to -1.
For each space-time combination, 30 observations are simulated.
}
\keyword{datasets}
