% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_data_set_up.R, R/dengue_data_set_up2.R,
%   R/dengue_data_set_up_version01.R
\name{dengue_data_set_up}
\alias{dengue_data_set_up}
\alias{dengue_data_set_up_2}
\title{This function creates the data needed to fit the basic model used in our paper.
It also outputs the precision matrix for the spatial ICAR model. We remove region 194,
it is disconnected. This function is based on "Combined effects of hydrometeorological hazards and urbanisation on dengue risk in Brazil: a spatiotemporal modelling study"}
\usage{
dengue_data_set_up(nt)

dengue_data_set_up_2(nt)

dengue_data_set_up(nt)
}
\description{
This function creates the data needed to fit the basic model used in our paper.
It also outputs the precision matrix for the spatial ICAR model. We remove region 194,
it is disconnected. This function is based on "Combined effects of hydrometeorological hazards and urbanisation on dengue risk in Brazil: a spatiotemporal modelling study"

This function creates the data needed to fit the basic model used in our paper.
It also outputs the precision matrix for the spatial ICAR model. We remove region 194,
it is disconnected. This function is based on "Combined effects of hydrometeorological hazards and urbanisation on dengue risk in Brazil: a spatiotemporal modelling study"

This function creates the data needed to fit the basic model used in our paper.
It also outputs the precision matrix for the spatial ICAR model. We remove region 194,
it is disconnected. This function is based on "Combined effects of hydrometeorological hazards and urbanisation on dengue risk in Brazil: a spatiotemporal modelling study"
}
